Rename `swap-dialog` to `swap`and move swap state pages to `happy` folder
