+++ src/renderer/components/swap/state/SwapStopAlert.tsx
