+++ src/renderer/components/swap/state/pages/happy/DownloadingBinaryPage.tsx
@@ -5,7 +5,7 @@ import {
   Typography,
 } from '@material-ui/core';
 import React from 'react';
-import { SwapStateDownloadingBinary } from '../../../../../models/store';
+import { SwapStateDownloadingBinary } from '../../../../../../models/store';
 
 function LinearProgressWithLabel({
   value,
