+++ src/renderer/components/swap/state/pages/happy/ProcessExitedPage.tsx
@@ -1,8 +1,8 @@
 import { Box, Button, makeStyles, Paper, Typography } from '@material-ui/core';
 import React from 'react';
 import { resetSwap } from 'store/features/swap/swapSlice';
-import { SwapStateProcessExited } from '../../../../../models/store';
-import { useAppDispatch, useAppSelector } from '../../../../../store/hooks';
+import { SwapStateProcessExited } from '../../../../../../models/store';
+import { useAppDispatch, useAppSelector } from '../../../../../../store/hooks';
 
 const useStyles = makeStyles((theme) => ({
   leftButton: {
