+++ src/renderer/components/swap/state/pages/happy/ReceivedQuotePage.tsx
@@ -1,6 +1,6 @@
 import { Box, CircularProgress, Typography } from '@material-ui/core';
 import React from 'react';
-import { SwapStateReceivedQuote } from '../../../../../models/store';
+import { SwapStateReceivedQuote } from '../../../../../../models/store';
 
 type ReceivedQuotePageProps = {
   state: SwapStateReceivedQuote;
