+++ src/renderer/components/swap/state/pages/happy/StartedPage.tsx
