+++ src/renderer/components/swap/state/pages/happy/WaitingForBitcoinDepositPage.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import FileCopyOutlinedIcon from '@material-ui/icons/FileCopyOutlined';
 import { clipboard } from 'electron';
 import BitcoinIcon from 'renderer/components/icons/BitcoinIcon';
-import { SwapStateWaitingForBtcDeposit } from '../../../../../models/store';
+import { SwapStateWaitingForBtcDeposit } from '../../../../../../models/store';
 
 const useStyles = makeStyles((theme) => ({
   depositAddressOuter: {
